requirements.txt
streamlit-authenticator==0.1.3
streamlit
pandas
scikit-image
streamlit_option_menu
opencv-python
tk
unidecode
mysql-connector-python